clean up code, add image_utils.cpp, add screenshot and QR code, update README
