+++ Demo/iOS/AICamera/AICamera/ImageRecognizer.swift
@@ -31,7 +31,3 @@ class ImageRecognizer {
     }
     
 }
-
-
-
-
