+++ Demo/iOS/AICamera/AICamera/SSDData.m
@@ -14,9 +14,6 @@
 - (id)init
 {
     self = [super init];
-    if (self) {
-        
-    }
     return self;
 }
