+++ Demo/iOS/AICamera/AICamera/SSDModel.swift
@@ -11,7 +11,6 @@ import Foundation
 enum SSDModel : String {
     case PascalMobileNet300 = "pascal_mobilenet_300_66.paddle"
     case FaceMobileNet160 = "face_mobilenet_160_91.paddle"
-    case PascalMobileNet160 = "mobilenet_ssd_160.paddle"
     case PascalVGG300 = "vgg_ssd_net.paddle"
     
     func normDimension() -> (Int32, Int32)
@@ -21,8 +20,6 @@ enum SSDModel : String {
             case .PascalMobileNet300:
                 return (300, 300)
             case .FaceMobileNet160:
-                return (160, 160)
-            case .PascalMobileNet160:
                 return (160, 160)
             case .PascalVGG300:
                 return (300, 300)
